warnings: set warning level to 2
Fix various warnings enabling the warning level
to 2 regarding the size

+++ lib/esefilereader.h
@@ -33,12 +33,12 @@ class ESEFileReader : public ESEReader {
 
   virtual bool prepare ();
 
-  virtual ESEBuffer getBuffer (uint32_t size);
+  virtual ESEBuffer getBuffer (size_t size);
   virtual size_t    streamSize () { return m_fileSize; }
   virtual bool      isEOS () { return m_bufferSize == 0 && m_readSize == streamSize (); }
 
   private:
-  uint32_t readFile (int32_t data_size, int32_t pos = 0, bool append = false);
+  size_t readFile (size_t data_size, int32_t pos = 0, bool append = false);
 
   std::ifstream m_file;
   std::string   m_fileName;
